added and working on seed.js
